自分の記事へのQiita推薦を禁止

+++ Vue/src/views/ArticleDetail.vue
@@ -3,14 +3,17 @@
     <v-snackbar v-model="isPostedArticleToQiita">
       Qiitaへ記事を投稿しました!!!
     </v-snackbar>
+    <v-snackbar v-model="nonValidUser" timeout="2000">
+      自分の記事にQiita推薦はできません
+    </v-snackbar>
     <v-row>
       <v-col class="hidden-xs-only hidden-sm-only" :md="mdPlacement.buttons">
         <v-row id="qiita_btn">
           <!-- Qiitaボタン -->
           <v-col
-              cols="12"
-              style="text-align: center; padding: 0"
-              class="green--text"
+            cols="12"
+            style="text-align: center; padding: 0"
+            class="green--text"
           >
             {{ article.qiitaRecommendPoint }}
           </v-col>
@@ -110,19 +113,19 @@
       <v-col cols="12" sm="12" :md="mdPlacement.editor">
         <span v-show="EditorIsOpen">
           <FeedbackEditor
-              class="sticky"
-              @closeEditor="closeEditor"
-              @postFeedback="postFeedback"
-              :feedback="propsFeedback"
+            class="sticky"
+            @closeEditor="closeEditor"
+            @postFeedback="postFeedback"
+            :feedback="propsFeedback"
           />
         </span>
         <span v-show="!EditorIsOpen">
           <v-btn
-              color="gray"
-              icon
-              large
-              @click="openNewEditor"
-              class="toggle_editor_btn"
+            color="gray"
+            icon
+            large
+            @click="openNewEditor"
+            class="toggle_editor_btn"
           >
             <v-icon>mdi-comment-plus</v-icon>
           </v-btn>
@@ -133,7 +136,7 @@
 </template>
 
 <script>
-import {mapActions} from "vuex";
+import { mapActions } from "vuex";
 import Article from "../components/article_detail/Article";
 import Feedbacks from "../components/article_detail/Feedbacks";
 import FeedbackEditor from "../components/article_detail/FeedbackEditor";
@@ -160,11 +163,15 @@ export default {
       propsFeedback: {},
       //snackbarに使用するメソッド
       isPostedArticleToQiita: false,
+      nonValidUser: false,
     };
   },
   computed: {
     slug() {
       return this.$route.params.articleId;
+    },
+    loginUser() {
+      return this.$store.state.auth.loginUser;
     },
     article() {
       return this.$store.state.article.article;
@@ -173,8 +180,8 @@ export default {
       return this.$store.state.article.article.feedbacks;
     },
     mdPlacement() {
-      if (this.EditorIsOpen) return {buttons: 1, article: 7, editor: 4};
-      return {buttons: 2, article: 8, editor: 2};
+      if (this.EditorIsOpen) return { buttons: 1, article: 7, editor: 4 };
+      return { buttons: 2, article: 8, editor: 2 };
     },
     apiToken() {
       return this.$store.getters["auth/apiToken"];
@@ -187,7 +194,7 @@ export default {
     },
   },
   watch: {
-    apiToken: function () {
+    apiToken: function() {
       this.fetchArticle(this.slug);
       this.fetchMyArticle(this.slug);
       this.fetchRecommend(this.slug);
@@ -249,18 +256,22 @@ export default {
         this.$store.dispatch("article/deleteMyArticle", this.myArticleId);
       } else {
         this.$store.dispatch(
-            "article/registerMyArticle",
-            this.article.articleId
+          "article/registerMyArticle",
+          this.article.articleId
         );
       }
     },
     toggleRecommend() {
+      if (this.loginUser.uid == this.article.postedUser.uid) {
+        this.nonValidUser = true;
+        return;
+      }
       if (this.recommendId) {
         this.$store.dispatch("article/deleteRecommend", this.recommendId);
       } else {
         this.$store.dispatch(
-            "article/registerRecommend",
-            this.article.articleId
+          "article/registerRecommend",
+          this.article.articleId
         );
       }
     },
